+++ CHANGELOG.md
@@ -4,8 +4,32 @@ All notable changes to the "templator" extension will be documented in this file
 
 ## [Unreleased] All the versions below are unreleased versions
 
-## 0.0.2 2025-03-03
+## [0.0.4] - 2025-03-10
+
+- Added file watcher for Cpp
+- Added templates for TypeScript
+- In the command palette, the language selection, and template selection options are now sorted alphabetically
+- Removed logger, added a dependency for logging
+- Improved documentation
+- Dependencies updated, security vulnerabilities fixed
+
+## [0.0.3] - 2025-03-10
+
+- Added file watcher, handles automatic ctx menu activation
+- Added templates for C#, TS, Cpp and Python
+- Code cleanup and refactoring
+- Internal commands are hidden from the command palette, only the context menu is available
+- Shortcuts added for file generation, and language selection: 'cs'
+- Dependencies updated
+
+## [0.0.2] 2025-03-03
 
 - The extension is now ready for testing with the C# programming language, 7 basic templates added to it
 - Namespace generation for C# is also ready for testing
 - Added 1 simple template for TS and also 1 for python
+
+## [0.0.1] - 2025-02-21
+
+- Started the project, added basic functionality for file generation.
+
+> Note: The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)